Move the logic for adding span links to span meta to spanpointers.rs
